create useCUDDecks
